+++ components/zcash_address/src/kind/unified/ivk.rs
@@ -1,10 +1,10 @@
 use alloc::vec::Vec;
-use core::convert::{TryFrom, TryInto};
+use core::convert::TryInto;
 use zcash_protocol::constants;
 
 use super::{
-    private::{SealedContainer, SealedItem},
-    Container, Encoding, ParseError, Typecode,
+    private::{SealedContainer, SealedDataItem},
+    Container, DataTypecode, Encoding, Item, ParseError,
 };
 
 /// The set of known IVKs for Unified IVKs.
@@ -46,31 +46,27 @@ pub enum Ivk {
     },
 }
 
-impl TryFrom<(u32, &[u8])> for Ivk {
-    type Error = ParseError;
-
-    fn try_from((typecode, data): (u32, &[u8])) -> Result<Self, Self::Error> {
+impl SealedDataItem for Ivk {
+    fn parse(typecode: DataTypecode, data: &[u8]) -> Result<Self, ParseError> {
         let data = data.to_vec();
-        match typecode.try_into()? {
-            Typecode::P2pkh => data.try_into().map(Ivk::P2pkh),
-            Typecode::P2sh => Err(data),
-            Typecode::Sapling => data.try_into().map(Ivk::Sapling),
-            Typecode::Orchard => data.try_into().map(Ivk::Orchard),
-            Typecode::Unknown(_) => Ok(Ivk::Unknown { typecode, data }),
+        match typecode {
+            DataTypecode::P2pkh => data.try_into().map(Ivk::P2pkh),
+            DataTypecode::P2sh => Err(data),
+            DataTypecode::Sapling => data.try_into().map(Ivk::Sapling),
+            DataTypecode::Orchard => data.try_into().map(Ivk::Orchard),
+            DataTypecode::Unknown(typecode) => Ok(Ivk::Unknown { typecode, data }),
         }
         .map_err(|e| {
-            ParseError::InvalidEncoding(format!("Invalid ivk for typecode {}: {:?}", typecode, e))
+            ParseError::InvalidEncoding(format!("Invalid ivk for typecode {:?}: {:?}", typecode, e))
         })
     }
-}
 
-impl SealedItem for Ivk {
-    fn typecode(&self) -> Typecode {
+    fn typecode(&self) -> DataTypecode {
         match self {
-            Ivk::P2pkh(_) => Typecode::P2pkh,
-            Ivk::Sapling(_) => Typecode::Sapling,
-            Ivk::Orchard(_) => Typecode::Orchard,
-            Ivk::Unknown { typecode, .. } => Typecode::Unknown(*typecode),
+            Ivk::P2pkh(_) => DataTypecode::P2pkh,
+            Ivk::Sapling(_) => DataTypecode::Sapling,
+            Ivk::Orchard(_) => DataTypecode::Orchard,
+            Ivk::Unknown { typecode, .. } => DataTypecode::Unknown(*typecode),
         }
     }
 
@@ -89,7 +85,7 @@ impl SealedItem for Ivk {
 /// # Examples
 ///
 /// ```
-/// use zcash_address::unified::{self, Container, Encoding};
+/// use zcash_address::unified::{self, Container, Encoding, Item};
 ///
 /// # #[cfg(not(feature = "std"))]
 /// # fn main() {}
@@ -100,28 +96,22 @@ impl SealedItem for Ivk {
 ///
 /// let (network, uivk) = unified::Uivk::decode(example_uivk)?;
 ///
-/// // We can obtain the pool-specific Incoming Viewing Keys for the UIVK in
-/// // preference order (the order in which wallets should prefer to use their
-/// // corresponding address receivers):
-/// let ivks: Vec<unified::Ivk> = uivk.items();
+/// // We can obtain the pool-specific Incoming Viewing Keys for the UIVK.
+/// let ivks: &[Item<unified::Ivk>] = uivk.items_as_parsed();
 ///
-/// // And we can create the UIVK from a list of IVKs:
-/// let new_uivk = unified::Uivk::try_from_items(ivks)?;
+/// // And we can create the UIVK from a vector of IVKs:
+/// let new_uivk = unified::Uivk::try_from_items(ivks.to_vec())?;
 /// assert_eq!(new_uivk, uivk);
 /// # Ok(())
 /// # }
 /// ```
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Uivk(pub(crate) Vec<Ivk>);
+pub struct Uivk(pub(crate) Vec<Item<Ivk>>);
 
 impl Container for Uivk {
-    type Item = Ivk;
+    type DataItem = Ivk;
 
-    /// Returns the IVKs contained within this UIVK, in the order they were
-    /// parsed from the string encoding.
-    ///
-    /// This API is for advanced usage; in most cases you should use `Uivk::items`.
-    fn items_as_parsed(&self) -> &[Ivk] {
+    fn items_as_parsed(&self) -> &[Item<Ivk>] {
         &self.0
     }
 }
@@ -146,7 +136,7 @@ impl SealedContainer for Uivk {
     /// The HRP for a Bech32m-encoded regtest Unified IVK.
     const REGTEST: &'static str = constants::regtest::HRP_UNIFIED_IVK;
 
-    fn from_inner(ivks: Vec<Self::Item>) -> Self {
+    fn from_inner(ivks: Vec<Item<Self::DataItem>>) -> Self {
         Self(ivks)
     }
 }
@@ -164,10 +154,10 @@ mod tests {
         sample::select,
     };
 
-    use super::{Ivk, ParseError, Typecode, Uivk};
-    use crate::kind::unified::{
-        private::{SealedContainer, SealedItem},
-        Container, Encoding,
+    use super::{Ivk, ParseError, Uivk};
+    use crate::{
+        kind::unified::{private::SealedContainer, Encoding as _},
+        unified::{Item, Typecode},
     };
     use zcash_protocol::consensus::NetworkType;
 
@@ -209,8 +199,8 @@ mod tests {
             shielded in arb_shielded_ivk(),
             transparent in prop::option::of(arb_transparent_ivk()),
         ) -> Uivk {
-            let mut items: Vec<_> = transparent.into_iter().chain(shielded).collect();
-            items.sort_unstable_by(Ivk::encoding_order);
+            let mut items: Vec<_> = transparent.into_iter().chain(shielded).map(Item::Data).collect();
+            items.sort_unstable_by(Item::encoding_order);
             Uivk(items)
         }
     }
@@ -306,11 +296,14 @@ mod tests {
     #[test]
     fn duplicate_typecode() {
         // Construct and serialize an invalid UIVK.
-        let uivk = Uivk(vec![Ivk::Sapling([1; 64]), Ivk::Sapling([2; 64])]);
+        let uivk = Uivk(vec![
+            Item::Data(Ivk::Sapling([1; 64])),
+            Item::Data(Ivk::Sapling([2; 64])),
+        ]);
         let encoded = uivk.encode(&NetworkType::Main);
         assert_eq!(
             Uivk::decode(&encoded),
-            Err(ParseError::DuplicateTypecode(Typecode::Sapling))
+            Err(ParseError::DuplicateTypecode(Typecode::SAPLING))
         );
     }
 
@@ -330,33 +323,5 @@ mod tests {
             Uivk::parse_internal(Uivk::MAINNET, &encoded[..]),
             Err(ParseError::OnlyTransparent)
         );
-    }
-
-    #[test]
-    fn ivks_are_sorted() {
-        // Construct a UIVK with ivks in an unsorted order.
-        let uivk = Uivk(vec![
-            Ivk::P2pkh([0; 65]),
-            Ivk::Orchard([0; 64]),
-            Ivk::Unknown {
-                typecode: 0xff,
-                data: vec![],
-            },
-            Ivk::Sapling([0; 64]),
-        ]);
-
-        // `Uivk::items` sorts the ivks in priority order.
-        assert_eq!(
-            uivk.items(),
-            vec![
-                Ivk::Orchard([0; 64]),
-                Ivk::Sapling([0; 64]),
-                Ivk::P2pkh([0; 65]),
-                Ivk::Unknown {
-                    typecode: 0xff,
-                    data: vec![],
-                },
-            ]
-        )
     }
 }